src/gather.c: Set GLOB_BRACE to 0 if it is not defined
GLOB_BRACE is a GNU extension that is not available on libc's such as
musl libc. To workaround that issue this commit introduces a fallback
that sets GLOB_BRACE to zero if it isn't defined. Beware that this does
not fix the lack of brace globbing on musl but instead just makes it
compile without this feature.